refactor: use local icons to fix failing tests
